+++ READMEs/marketplace-flow.md
@@ -170,7 +170,7 @@ did = asset.did  # did contains the datatoken address
 
 #Mint the datatokens
 from decimal import Decimal
-from ocean_lib.web3_internal.utils import to_wei
+from ocean_lib.web3_internal.currency import to_wei
 data_token.mint(alice_wallet.address, to_wei(Decimal("100.0")), alice_wallet)
 
 #In the create() step below, Alice needs ganache OCEAN. Ensure she has it.
