+++ tests/integration/test_compute_flow.py
@@ -12,7 +12,7 @@ from ocean_lib.models.data_token import DataToken
 from ocean_lib.models.metadata import MetadataContract
 from ocean_lib.ocean.util import get_contracts_addresses
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
-from ocean_lib.web3_internal.utils import to_wei
+from ocean_lib.web3_internal.currency import to_wei
 from tests.resources.ddo_helpers import (
     get_algorithm_meta,
     get_registered_algorithm_ddo,
